Rewrite annotation layers so they don't occlude function metadata
With the original approach, wrapped APIs would report their arity
as (arity-at-least 0) even if they had more specifically known
arities. This was because the APIs were being wrapped with new
functions that accepted an arbitrary number of arguments and simply
forwarded them to the wrapped functions. In order to preserve the
function metadata, we now use macros that simply write out functions
with the same signature as the original functions being wrapped, using
a straightforward way to indicate what that signature is.